refactor(console): 修复rich控制台问题，添加控制台模块
