+++ src/ErisPulse/utils/__init__.py
@@ -7,9 +7,11 @@ ErisPulse SDK 工具模块
 from .package_manager import PackageManager
 from .reload_handler import ReloadHandler
 from .cli import CLI
+from .console import console
 
 __all__ = [
     "PackageManager",
     "ReloadHandler",
     "CLI",
+    "console",
 ]
